+++ cognite/dm_clients/bin/schema_render.sh
@@ -7,11 +7,4 @@ set -e
 SCHEMA_FILE=$(cat "$CONFIG" | extract 'schema_file')
 SCHEMA_MODULE=$(cat "$CONFIG" | extract 'schema_module')
 
-cat <<'EOF' > "$SCHEMA_FILE"
-# THIS FILE IS AUTO-GENERATED!
-# Use `dm_clients schema render` to update it, see `dm_clients --help` for more information.
-
-
-EOF
-
 python -m "$SCHEMA_MODULE" >> "$SCHEMA_FILE"