+++ examples/cinematography_domain/usage.py
@@ -5,7 +5,6 @@ from typing import List, cast
 
 from cinematography_domain.client import CineClient, get_cine_client
 from cinematography_domain.schema import Movie, Person
-
 from cognite.dm_clients.custom_types import JSONObject, Timestamp