+++ CHANGELOG.md
@@ -13,6 +13,24 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.2.0] - 19-04-23
+
+### Added
+
+* Support for converting from `.graphql` to `pydantic`. This means you can now write your types in a `.graphql` file
+  and automatically generate the `pytandic` counterpart.
+* CLI for doing the conversion between GraphQL and `pydantic`. The following two commands
+  * `dm togql` converts a pydantic schema to `.graphql` and creates the `client.py` which is the SDK to interact
+    with the data model through Python.
+  * `dm topython` converts a `.graphql` to `pytandic` classes. This command also creates the `client.py`, i.e., the
+    SDK to interact with the data model through Pytho
+* Documentation of motivation and usage of package.
+
+### Changed
+
+* CLI command going from `gqlpygen` to `pygen`. This is for ease of use.
+
+
 ## [0.1.4]
 
 Changed configuration / settings from a static `config.yaml` to dynaconf and `settings.toml`.